refac #20: api route 정리
